add prompt model section
